Moderation: consolidate all moderation into a single /moderation console
Unify the moderation surfaces into one moderator dashboard and
add a calendar moderation tab.

Calendar moderation tab (new):
- Events | Proposals segmented views. Events get edit + hide/unhide;
  proposals get a Pending/Approved/Rejected status filter with
  approve/reject on pending ones.
- Extract the chain-event editor (create/edit form modal + hide toggle)
  from the calendar page into a shared useChainEventEditor hook, reused by
  the public calendar and the moderation console (no duplication).
- hooks: useCalendarProposals returns all statuses; add
  useModeratableEvents (all on-chain events incl. hidden, one row per id).

/moderation console:
- Rename AdminConsole -> ModerationConsole and move it from /inbox/admin to
  /moderation (wrapped in RequireModerator), with all five tabs: Channels,
  Poll comments, Token comments, Calendar, Moderators.
- Remove the duplicate moderation tabs (channels, poll/token comments,
  calendar, moderators) from /dao, leaving it pure governance.
- Route the menu link and the route guard through one mock-aware role path
  (useGateRoles / useCanModerate), so menu, guard and role resolution agree.
