+++ src/data/calendar-chain/hooks.ts
@@ -5,8 +5,13 @@ import useSWR from 'swr'
 import { useUserStore } from '@context/userStore'
 import { useGateRoles } from '@data/roles'
 import { fetchProposals } from './api'
+import { fetchChainCalendarEvents } from './index'
 import { fetchEventContent } from './ipfs'
-import type { CalendarProposal, CalendarEventContent } from './types'
+import type {
+  CalendarProposal,
+  CalendarEventContent,
+  CalendarFeedEvent,
+} from './types'
 
 export interface CalendarRoles {
   /** Connected wallet address, for own-event ("proposed by me") checks. */
@@ -41,17 +46,16 @@ export interface ProposalListItem extends CalendarProposal {
 }
 
 /**
- * Pending proposals with their proposed document resolved from IPFS, for the
- * moderator queue — moderators must see what they approve, not just a title.
- * Only fetches when `enabled` (i.e. the viewer can moderate).
+ * All community proposals (any status), each with its proposed document
+ * resolved from IPFS — moderators must see what they approve, not just a title.
+ * Callers filter by status (pending / approved / rejected). Only fetches when
+ * `enabled` (i.e. the viewer can moderate).
  */
 export function useCalendarProposals(enabled: boolean) {
   return useSWR<ProposalListItem[]>(
     enabled ? 'calendar/proposals' : null,
     async () => {
-      const proposals = (await fetchProposals()).filter(
-        (p) => p.status === 'pending'
-      )
+      const proposals = await fetchProposals()
       const docs = await Promise.allSettled(
         proposals.map((p) => fetchEventContent(p.proposedCid))
       )
@@ -64,6 +68,34 @@ export function useCalendarProposals(enabled: boolean) {
           doc,
         }
       })
+    },
+    { revalidateOnFocus: false, dedupingInterval: 15_000 }
+  )
+}
+
+/**
+ * All on-chain events for the moderation console, collapsed to one row per
+ * series (recurring occurrences share an eventId). Includes hidden events so a
+ * moderator can unhide them. Only fetches when `enabled` (viewer can moderate).
+ */
+export function useModeratableEvents(enabled: boolean) {
+  const address = useUserStore((st) => st.address)
+  return useSWR<CalendarFeedEvent[]>(
+    enabled ? ['calendar/mod-events', address] : null,
+    async () => {
+      const events = await fetchChainCalendarEvents({
+        includeHidden: true,
+        viewerAddress: address,
+      })
+      // Collapse recurring occurrences to one entry per series (by eventId).
+      const seen = new Set<number>()
+      const series: CalendarFeedEvent[] = []
+      for (const ev of events) {
+        if (seen.has(ev.eventId)) continue
+        seen.add(ev.eventId)
+        series.push(ev)
+      }
+      return series.sort((a, b) => a.eventId - b.eventId)
     },
     { revalidateOnFocus: false, dedupingInterval: 15_000 }
   )
