+++ src/index.jsx
@@ -22,7 +22,6 @@ import {
   DaoParameters,
   DaoProposals,
   SubmitDaoProposals,
-  DaoModerators,
   DaoFees,
 } from '@pages/dao/tabs'
 import {
@@ -93,12 +92,7 @@ import ChannelView from '@components/channels/ChannelView'
 import CreateChannel from '@components/channels/CreateChannel'
 import ChannelSettings from '@components/channels/ChannelSettings'
 import DmRedirect from '@components/channels/DmRedirect'
-import {
-  AdminConsole,
-  ChannelsAdmin,
-  CommentsAdmin,
-  RequireModerator,
-} from '@pages/admin'
+import { ModerationConsole, RequireModerator } from '@pages/admin'
 import { StatsPage } from '@pages/stats'
 
 import Sync from '@pages/sync'
@@ -240,32 +234,7 @@ const router = createBrowserRouter(
         <Route path="proposals" element={<DaoProposals />} />
         <Route path="submit" element={<SubmitDaoProposals />} />
         <Route path="stats" element={<StatsPage />} />
-        <Route path="moderators" element={<DaoModerators />} />
         <Route path="fees" element={<DaoFees />} />
-        <Route
-          path="channels"
-          element={
-            <RequireModerator>
-              <ChannelsAdmin />
-            </RequireModerator>
-          }
-        />
-        <Route
-          path="poll-comments"
-          element={
-            <RequireModerator>
-              <CommentsAdmin kind="poll" />
-            </RequireModerator>
-          }
-        />
-        <Route
-          path="token-comments"
-          element={
-            <RequireModerator>
-              <CommentsAdmin kind="token" />
-            </RequireModerator>
-          }
-        />
         <Route path="*" element={<DaoParameters />} />
       </Route>
       <Route path="proposal/:id" element={<ProposalDisplay />} />
@@ -292,7 +261,14 @@ const router = createBrowserRouter(
       <Route path="inbox/channels/:id" element={<ChannelView />} />
       <Route path="inbox/channels/:id/settings" element={<ChannelSettings />} />
       <Route path="inbox/dm/:address" element={<DmRedirect />} />
-      <Route path="inbox/admin" element={<AdminConsole />} />
+      <Route
+        path="moderation"
+        element={
+          <RequireModerator>
+            <ModerationConsole />
+          </RequireModerator>
+        }
+      />
       <Route path="activity" element={<TeiaActivity />} />
       <Route path="tags/:tag" element={<Tags />} />
       <Route path="tz/:address/*" element={<Display />}>
