+++ src/pages/calendar/index.module.scss
@@ -157,73 +157,6 @@ details[open] > .subscribe_toggle::after {
   margin-bottom: var(--spacing, 1rem);
 }
 
-.form_overlay {
-  position: fixed;
-  inset: 0;
-  z-index: 900;
-  background: rgba(0, 0, 0, 0.65);
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  overflow-y: auto;
-  padding: max(5vh, 96px) 16px 5vh;
-  box-sizing: border-box;
-}
-
-.form_modal {
-  background: var(--background-color);
-  color: var(--text-color);
-  border: 1px solid var(--border-color);
-  width: min(720px, 92vw);
-  max-height: 90vh;
-  overflow-y: auto;
-
-  // EventForm already draws its own border/padding; avoid doubling up.
-  form {
-    border: none;
-    margin-bottom: 0;
-  }
-}
-
-// Moderator proposal queue.
-.queue {
-  margin-bottom: var(--spacing, 1rem);
-  border: 1px solid var(--border-color);
-  padding: var(--spacing, 1rem);
-}
-
-.queue_title {
-  font-size: $font-large;
-  margin: 0 0 var(--spacing, 1rem);
-}
-
-// One proposal: header line (kind, proposer, actions) + the proposed event card.
-.queue_item {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 12px 0;
-  border-top: 1px solid var(--border-color);
-}
-
-.queue_head {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: var(--spacing, 1rem);
-  flex-wrap: wrap;
-}
-
-.queue_meta {
-  color: var(--text-secondary);
-  font-size: var(--font-size-small);
-}
-
-.queue_actions {
-  display: flex;
-  gap: 8px;
-}
-
 .upcoming {
   margin: var(--spacing, 1rem) 0 2rem;
 }
